refactor(engine): extract dmn transformer
